List files to download from DNOs
This adds dagster assets to build lists of available files, the list of
downloaded files and then from that, the list of new files we need to
download.

This is a bit of a mammoth commit as I spent a lot of time figuring out
the best asset structure and how to test it. It's probably not perfect
but some decisions I made along the way to make it better were:
- Keeping things separated by DNO so that if one's API is down, we can
  still download from the others
- Using Dagster's implicit IO Manager setup to pass data around between
  assets, rather than persisting stuff anywhere specific inside the
  asset definition - it's small data and we don't really want to keep it
  around, plus this is easily portable/configurable later (e.g. to use
  tmp files in S3 instead of memory/files locally)
- Using requests for http and responses for mocking in tests over the
  more complex async stuff I experimented with earlier
- Using fsspec to abstract filesystems rather than complex Dagster
  resources - I couldn't figure out a good way to match Dagster's YAML
  config with code that was file vs S3 agnostic

One thing I know I still need to figure out is how we manage secrets in
the config and whether we can commit that config somewhere. I'll tackle
that when I come to put this in production.

Since I dropped aiohttp I also updated the ADR that used it to use
requests instead.

Closes https://github.com/orgs/centre-for-ai-and-climate/projects/1/views/1?pane=issue&itemId=80869231