Switch to more Dagster-native sensors (for SSEN)
Starting the next step of the pipeline - downloading files - made me
realise that the way I'd done the first steps was not going to be very
useful.

We want Dagster to manage the downloading of lots of files - retrying
failures, managing concurrency, etc. Their way of doing that is with
partitioned assets, which then in turn rely on sensors to find the
partitions and trigger materializations. I'd essentially built my own
ad-hoc sensors by modelling various interim steps as assets, which then
meant I couldn't easily create partitioned assets.

Our assets were also set up to need runtime configuation. This was
intended to make them easy to change between dev and prod, but also had
the side-effect of meaning we *had* to configure them in YAML, which was
repeated for every single asset. A lot of the interim assets I'd made
were actually reusing the same config, so it was quite annoying to setup
and then the config was just stored in Dagster's DB, rather than version
controlled.

When looking at using sensors, I also realised that they would need to
supply the config to the asset every time they ran, which would be
another big complication.

Therefore, I've basically completely rewritten the pipeline, so that we
have:
- A Resource encapsulating the SSEN API
- A Resource modelling the place raw files are stored
- A Sensor that finds the files we want to download
- A Partitioned Asset that represents the files we want to download
  (from SSEN only, initially)

With this, Dagster also manages remembering which files have been
downloaded (the "cursor") for us, and gives us a nice UI to backfill,
refresh, downloading specific subsets, etc.

Our resources are now configured in code, so we can version control the
setup and run it in different environments.

Overall, it's way less code, better encapsulated and easier to work
with.

I've only done SSEN so far, as a way to work through the approach
before replicating it for the other sources.