Add monthly parquet aggregations of SSEN data
The next step after getting raw data is to start to shape it to be more
how we want it. We'll do everything in monthly partitions for the
moment, so it makes sense to partition the raw data that way too.

I decided the best thing to do would be to use parquet files internally
as well, to minimise storage and transfer sizes, so this adds an asset
which produces those for SSEN. The only really tricky bit is triggering
their creation because the raw data is dynamically partitioned. So
instead of any native Dagster magic, we need our own sensor to watch for
materialisations of the raw data assets.