Run monthly parquet sensor only every 5 minutes
Trying to avoid a situation where we could have multiple runs of the
job competing to update the monthly parquet files at the same time,
e.g. when a big backfill is running. 5 mins should be enough time for
a large number of monthly files to be written.

This isn't perfect, during a backfill we're still going to generate
files multiple times as we catch up with the new/updated raw partitions,
but that seems better (for now) than the alternatives of manually
scheduling things in sequence, or using complex idempotency/concurrency
controls.