Add SSEN substation locations ADR
In our prototype we added locations to SSEN data, because it was missing
and we thought it was really important. We wanted to spend more time
checking the accuracy of that approach and try to improve it. Doing so
turned out to be quite the boondoggle, so I've documented the approach
in an ADR for future reference.