Add SSEN substation location lookup from ONSPD
This implements part of ADR 0003, which is to use the ONS Postcode
Directory as our backstop of substation locations, but with some quite
heavy filtering to remove locations that are not trustworthy enough to
depend on.

In building this, I realised I was blurring the lines between the Client
and Asset classes I'm building, particularly around the methods that
load raw data as data frames. I've attempted to clarify that a bit,
which resulted in me removing the ONSPD parquet asset I'd added
previously.

The general approach I've settled on is of having the client classes be
responsible for opening and interpretting the data, but not selecting
subsets or otherwise cleansing it.

I've also slightly tweaked the raw data files so that they report the
schema and other metadata in the same way as the derived data files do,
including fixing some metadata for the LV feeder lookup file which we
now know is incorrect.