Make gzipping downloaded SSEN files optional
We need to download other files from SSEN's API that are already
compressed, e.g. as zip files, so we need an option to download things
as-is as well.

This is starting to feel like too low-level a concern for the API
clients and something which I should perhaps handle in a more generic
way, but I'm following the rule of three at the moment (and I only have
two examples).