Add SSEN substation locations to monthly parquet files
This commit uses the new substation location lookups to add locations to
all the SSEN data when processing the daily files.

I had to make a few changes to make this possible:

- I've simplified the schema to remove the dictionary columns. I wanted
  to use dictionaries to save space, but because pyarrow is very limited
  in what it can join when joining tables, I had to remove them.
- I've added a new column to the schema for the substation location,
  which is currently modelled on the LV feeder data schema (i.e. a
  simple string with "lat, lon")
- I've added the lookup on a daily file basis - one month's data is too
  large to fit in memory so we have to do it as we go.

This last point has some interesting implications for the future. We
might have to do all our processing on a daily basis if we need to load
the whole table into memory - i.e. to use pandas/geopandas.