Add a more robust method to get OutputFile's paths
We want to record this in Dagsters metadata because otherwise it's hard
to know where files have gone, but the `name` attribute of the file I
was using is not portable to S3.

This still feels like a smell - the location should be more
authoritatively chosen, or not matter at all. But for now, this is a
fix.