Move PyProj grid shift check to top of the module
We need to make sure these transformation grids are available for any
function that does BNG -> WGS84 conversion, so I've made this a
top-level assertion.

After testing on Dagster cloud, the method which works locally to
manually download the grid shift files doesn't work, but enabling
network access for PyProj does, which simplifies the check as well.