Hide dependency between dynamic -> monthly partitions from Dagster
This is a workaround for a bug (maybe) introduced in https://github.com/dagster-io/dagster/pull/25107
where Dagster started checking partition keys requested in backfills
were valid. In our case, where we're depending on a dynamically
partitioned asset from a monthly partitioned one, this check fails. I
think this is basically because the dynamic asset doesn't have matching
monthly partition keys, but I haven't looked into it in detail yet.

This worked previously in 1.8.8 where this check wasn't present, and
"hiding" the dependency from Dagster like this seems to work around the
issue for now.