Add all partition mapping to lv feeder monthly parquet
Rather than removing the dependency completely, as I did in #6, this
tells Dagster a slightly less untrue thing - that each month depends on
the whole set of raw data partitions. This isn't right either, but it
allows Dagster to check that things are somewhat "valid" and thus allows
us to run backfills, both manually and automatically.

It seems like mapping dynamic partitions is a bit of a pain point in
Dagster, that they don't have a great solution for (https://github.com/dagster-io/dagster/issues/13139)
so we're probably going against the grain here.

Note: in this case I could kinda cheat by creating a
StaticPartitionMapping, because each filename is a day, or I could even
maybe represent the raw data as a DailyPartitionsDefinition, but the
fact that the latest data is released ad-hoc would spoil it. That
solution also wouldn't work for NGED, where the data is released in
random part chunks, so I haven't done it here. It might be worth seeing
if we could better represent that aspect of the data in the future,
though.