Make SSEN LV Feeder downloads more robust to bad files
We've just had our first experience of a bad file from the SSEN API.
They have published a 0B file in place of the data for 11-18-2024.

With our existing pipeline, this breaks the whole of November's data,
because we download it without any checks, and then the monthly
partition falls over when it tries to parse it.

To fix this, I've added a check to the raw file step which tries to
parse the data into a PyArrow table and if it fails, deletes the file.
This now raises an error at the earlier stage, so we see a more useful
failure upstream, but it also means we don't download a bad raw file, so
we can continue to produce the monthly data without interruption.

Perhaps in the future we'll want to introduce a quarantine step or
similar, but this is okay for now.

I took this opportunity to refactor the PyArrow table creation into the
SSEN API client, so it lives more closely to all the other code that
knows about us GZipping the files.

I also took this opportunity to add some metadata to the
materializations, because it was missing from my first pass at these
files.