Fix #18: Handle missing inputs in monthly parquet
Dagster creates monthly assets as soon as the month starts, but the
daily inputs lag behind a bit - often we don't have the data for the 1st
until the 5th or 6th. Before this change, we would output an empty file
for the monthly raw data, then the partition run for the final output
file would crash when reading that empty file.

This commit makes two changes to handle this situation:
1. Stop outputting empty files for the monthly raw data
2. Catch errors in the combined output asset so we're robust to missing
   inputs

Both of these are made a bit more complex by the fact that we're
streaming data into files, so we open them before we find out things are
missing. This means we need to delete files to tidy up after ourselves.

This is another pointer that the way we're linking our dynamic raw files
to monthly partitions is a bit fragile, but I'm not sure what to do
about it.