Add tests for automation (and fix it too)
I wanted to add some tests to the automation code, because I suspected
we were using it slightly wrong and it's hard to test otherwise - you
can't even run it in branch deploys.

It took quite a while to figure out how to test things. In particular
realising that you need to run a tick of the automation conditions every
time something changes, not just at the point you want changes to be
picked up, because it's like hand-cranking a state machine with the way
the cursor is passed between steps. It still feels a bit low-level,
because the test helpers Dagster provides are quite basic, however,
there are at least some tests which cover the basic functionality we
expect.

In doing this, I realised that the built-in conditions we were using to
automate stuff weren't nuanced enough. Rather than have a lot of inline
modifications as some of their guides suggest, I've added new custom
conditions which are more specific to our needs. I think this is a good
starting point and we will probably add more as we go along.