Hotfix automation of combined geoparquet
I'm working on a more thorough set of tests for these automation conditions, but I've already realised this is set up wrong. on_missing only materialises the asset if it's totally new, but we're updating the upstream assets all the time and want that reflected, so we want any_deps_updated instead.