Download ssen postcode mapping only when it changes
Now that we've fixed our automation, we'll be triggering updates to the
entire dataset every day, regardless of whether anything has actually
changed, which is a bit wasteful.

I'm not sure this will be a huge improvement, as this mapping file seems
to update quite often, but it might save us a few run minutes in the
short term.

Closes #22