Add a sensor for NGED's raw LV Feeder data
This is the first step in downloading NGED data - working out what files
need to be downloaded and creating new dynamic partitions for the
related asset. The actual asset is a no-op for now, I will flesh it out
later.

I've copied the approach I took with SSEN - adding an API client which
hits NGED's API to get a list of available files. This is slightly
different though because NGED use their CKAN instance to hold their
data, rather than a dedicated API. Note this is a bit of a copy/paste
job (e.g. the download_file function is identical to SSEN's) but I'll
refactor that later.

Because of the need to use a sensor cursor based on file created times,
I've also extended the AvailableFile model to include an (optional)
created datetime.