Flesh out asset to download NGED LV feeder files
Now the asset we previously stubbed out can actually download the data.

In the process I realised NGED capitalises some words in the column
names, meaning we can't reuse exactly the same pyarrow schema for
processing the data. I've moved the schemas back into the API clients
for now.

The api clients are still duplicating a lot of the downloading code,
with the minor difference that NGED needs authorisation headers.