Add an asset and sensor for NGED LV feeder monthly parquet
This mostly copies the SSEN approach, but it's a lot simpler because the
data doesn't really have any major holes that need filling. The main
complexity is in figuring out exactly which raw files map to a specific
monthly partition.

As part of this I've messed again with the pyarrow schemas for the
parquet files. Now I have one in core.py which represents the data
from DNOs as it should be, and one in each API Client which represents
how they actually deliver it. This distinction is moot for SSEN but not
for NGED, so it seems a more sensible way to do things.

I've also refactored a few shared things in the sensors and CSV ->
Parquet process which were easy - there's more we could do here when we
have more examples I expect.

Note this increases the number of places we do
instance.report_runless_asset_event because it's a simpler way to do
testing than actually materialising the asset. Hopefully this stops
being experimental soon.