Extract helper to query dagster materialisations
We were doing something similar in multiple places, some better than
others because they dealt with the pagination of the underlying API
rather than trying to ignore. I was OK with that, but then realised
there was a mismatch with dagster clouds limits, so it was better to
fix it properly and put everything in one place to do that.