Add NGED to combined geoparquet, sorted correctly
This adds NGED data to the combined geoparquet output. In line with
ADR 0005, I've changed how we process this in batches so that we can get
output data which is sorted by data_collection_log_timestamp first, i.e.
different DNOs data is interleaved over each settlement period. This
matches how we did it in the alpha prototype, but is not how we were
doing it with SSEN due to misunderstanding.

This does not completely solve the problem, because I still need to go
back to the staging files and make sure _they_ are sorted correctly for
us to be able to efficiently filter them here.

I also need to apply some NGED specific fixes at that stage, because
refactoring this code to work for multiple DNOs made it obvious to me
that that's where such fixes should be applied.