Update automation for geoparquet now we have two dependencies
Now that we have two dependencies, we can't have a condition that no
dependencies must be missing, because DNOs publish data at different
rates and so there will legitimately be some missing for a while
(potentially a couple of months).

This adds a new automation condition specifically for the geoparquet
which captures that requirement.

Note: this has resulted in a change to a test of the automation - for
some reason, triggering the ssen monthly files by incrementally building
its dependencies does not then trigger the geoparquet like it did
before. I was somewhat surprised that it did before (we didn't
materialise it, only trigger it) and I have another (passing) test which
tests the geoparquet automation directly, so I've modified the test for
now to match reality. I'm going to observe in prod though to make sure
it's working there.