Process SSEN raw files in daily batches
I switched these from byte-range-sized batches to all in one go to match
NGED when I realised I needed to start sorting them for easy retrieval
in daily batches for the downstream combined geoparquet asset. I did
this based on looking at the memory usage of the 2024-02 partition and
reasoning it wasn't _too_ big. I'd forgotten that partition only has
half the data in it, because SSEN started publishing on 2024-02-12.
Because of that, they can't run on Dagster any more because it runs out
of memory.

This fixes that, but keeps the ability to sort the data by batching
each daily file. Thankfully SSEN releases their data in this way,
otherwise we'd be a bit stuck.