Experiment: use Polars to streaming sort parquet
After the issues I had with SSEN's data being too big, now I've hit
similar problems with a few months of NGED data. This is harder to solve
because it doesn't naturally split into time partitions like SSEN does.

I hackily implemented my own approach to sorting the data in the
parquet file, by writing it out unsorted and then reading it back in
daily batches with filters, but it was a lot of code and pretty fragile.

This is an alternative using the Polars library, which claims to be able
to sort in a streaming way. It has one drawback in that it can't set the
parquet sorting_columns metadata, but I'd like to see whether it solves
the problem first.