Use local temp storage in combined geoparquet
Adds some temporary files, which we can use to cache both inputs and
outputs in order to speed up processing, as well as to provide a safety
net in case of failures during asset runs.

I initially tried this with another OutputFilesResource, but it was a
pain to configure the temp directory up front. In general, this has made
me realised the OutputFilesResource is a bit of a leaky abstraction
generally and I might have been better off just using fsspec directly.

Fixes #40 and #41